requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-dotenv
azure-storage-blob
azure-ai-documentintelligence==1.0.0b2
azure-cosmos==4.5.1
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.3.1
azure-eventhub==5.11.4
azure-storage-file-datalake==12.14.0
openai==1.35.13
Flask
flask-cors
langchain
langchain-openai==0.1.15
azure-identity